+++ R/frequentist.R
@@ -92,13 +92,14 @@ gp_lrs <- function(x) {
   n <- length(x)                             # sample size
   x <- sort(x)                               # put data in ascending order
   q0 <- 1 / (n + 1)
-  q2 <- n / (n+1)                            # for sample minimum and maximum
+  q2 <- n / (n + 1)                          # for sample minimum and maximum
   a <- sqrt((1 - q2) / (1 - q0))
   q1 <- 1 - a *(1 - q0)                      # `middle' quantile
   n0 <- 1
   n1 <- round((n + 1) * q1)
   n2 <- n                                    # corresponding order statistics
-  ns <- c(n0,n1,n2); qs <- c(q0,q1,q2)
+  ns <- c(n0,n1,n2)
+  qs <- c(q0,q1,q2)
   xs <- x[ns]
   r_hat <- (xs[3] - xs[2]) / (xs[2] - xs[1])
   xi_hat <- -log(r_hat) / log(a)
@@ -709,8 +710,8 @@ grimshaw_gp_mle <- function(x) {
 ###############################################################################
 
 gev_mle <- function(init, ...){
-    x <- stats::optim(init, gev_loglik, control = list(fnscale = -1),
-               hessian = FALSE, ...)
+    x <- stats::optim(init, gev_loglik, ..., control = list(fnscale = -1),
+               hessian = FALSE)
     temp <- list()
     temp$mle <- x$par
     temp$nllh <- -x$value
@@ -859,8 +860,8 @@ gev_lrs <- function(x){
 ###############################################################################
 
 os_mle <- function(init, gumbel = FALSE, ...){
-  x <- stats::optim(init, os_loglik, control = list(fnscale = -1),
-                    hessian = FALSE, gumbel = gumbel, ...)
+  x <- stats::optim(init, os_loglik, ..., control = list(fnscale = -1),
+                    hessian = FALSE, gumbel = gumbel)
   temp <- list()
   temp$mle <- x$par
   if (gumbel) {