+++ tests/testthat/test-rpost_bingp.R
@@ -1,4 +1,4 @@
-context("GP: rpost_rcpp vs rpost")
+context("binGP: rpost_rcpp vs rpost")
 
 # We check that the values simulated using rpost() and rpost_rcpp() are
 # (close enough to) identical when they are called using the same data,
@@ -10,9 +10,9 @@ my_tol <- 1e-5
 # Set a threshold.
 u <- quantile(gom, probs = 0.65)
 
-gp_test <- function(seed = 47, prior, n = 5, rotate = TRUE,
-                    trans = "none", use_phi_map = FALSE, data = gom,
-                    nrep = 2, ...){
+bingp_test <- function(seed = 47, prior, n = 5, rotate = TRUE,
+                       trans = "none", use_phi_map = FALSE, data = gom,
+                       nrep = 2, ...){
   if (prior == "user") {
     prior_rfn <- set_prior(prior = "flat", model = "gp", min_xi = -1)
     ptr_gp_flat <- create_prior_xptr("gp_flat")
@@ -22,11 +22,11 @@ gp_test <- function(seed = 47, prior, n = 5, rotate = TRUE,
     prior_cfn <- set_prior(prior = prior, model = "gp", ...)
   }
   set.seed(seed)
-  res1 <- rpost(n = n, model = "gp", prior = prior_rfn, thresh = u,
+  res1 <- rpost(n = n, model = "bingp", prior = prior_rfn, thresh = u,
                 data = data, rotate = rotate, trans = trans,
                 use_phi_map = use_phi_map, nrep = nrep)
   set.seed(seed)
-  res2 <- rpost_rcpp(n = n, model = "gp", prior = prior_cfn, thresh = u,
+  res2 <- rpost_rcpp(n = n, model = "bingp", prior = prior_cfn, thresh = u,
                      data = data, rotate = rotate, trans = trans,
                      use_phi_map = use_phi_map, nrep = nrep)
   return(list(sim1 = as.numeric(res1$sim_vals),
@@ -57,25 +57,25 @@ for (rotate in rotate_vals) {
     for (use_phi_map in use_phi_map_vals) {
       test_string <- paste("rotate =", rotate, "trans =", trans,
                             "use_phi_map =", use_phi_map)
-      x <- gp_test(prior = "flat", min_xi = -1,
+      x <- bingp_test(prior = "flat", min_xi = -1,
                    rotate = rotate, trans = trans, use_phi_map = use_phi_map)
       test_function(x, test_string)
-      x <- gp_test(prior = "flatflat", min_xi = -1,
+      x <- bingp_test(prior = "flatflat", min_xi = -1,
                    rotate = rotate, trans = trans, use_phi_map = use_phi_map)
       test_function(x, test_string)
-      x <- gp_test(prior = "norm", mean = c(0,0), cov = diag(100,2),
+      x <- bingp_test(prior = "norm", mean = c(0,0), cov = diag(100,2),
                    rotate = rotate, trans = trans, use_phi_map = use_phi_map)
       test_function(x, test_string)
-      x <- gp_test(prior = "mdi",
+      x <- bingp_test(prior = "mdi",
                    rotate = rotate, trans = trans, use_phi_map = use_phi_map)
       test_function(x, test_string)
-      x <- gp_test(prior = "jeffreys",
+      x <- bingp_test(prior = "jeffreys",
                    rotate = rotate, trans = trans, use_phi_map = use_phi_map)
       test_function(x, test_string)
-      x <- gp_test(prior = "beta",
+      x <- bingp_test(prior = "beta",
                    rotate = rotate, trans = trans, use_phi_map = use_phi_map)
       test_function(x, test_string)
-      x <- gp_test(prior = "user",
+      x <- bingp_test(prior = "user",
                    rotate = rotate, trans = trans, use_phi_map = use_phi_map)
       test_function(x, test_string)
     }